Adding script to search for strcp in a file or tree.
Adding script that if the argument is a file it will grep the file for
instances of strncpy and return the path to the file, and matching line
numbers. If the argument is a directory it will recursively find '.c'
files and grep them for instances of strcpy and report the path to the
file, and matching line numbers.